+++ 2D Dogfight Project/Assets/Scenes/Game.unity
@@ -495,7 +495,6 @@ MonoBehaviour:
   hand: {fileID: 0}
   _card: {fileID: 0}
   planeOutline: {fileID: 830315315}
-  _Preview: 0
 --- !u!114 &577942722
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -801,6 +800,49 @@ Transform:
     type: 3}
   m_PrefabInstance: {fileID: 1656602690}
   m_PrefabAsset: {fileID: 0}
+--- !u!1 &912275577
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 912275579}
+  - component: {fileID: 912275578}
+  m_Layer: 0
+  m_Name: CommandInvoker
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &912275578
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 912275577}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: afc8d97669a1c2e49a455844e707af46, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!4 &912275579
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 912275577}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 3.5097842, y: 0.7783298, z: -0.8359375}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 5
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1001 &991540772
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -1164,7 +1206,6 @@ MonoBehaviour:
   hand: {fileID: 0}
   _card: {fileID: 0}
   planeOutline: {fileID: 568179466}
-  _Preview: 0
 --- !u!1 &1318853549
 GameObject:
   m_ObjectHideFlags: 0
@@ -1341,7 +1382,6 @@ MonoBehaviour:
   hand: {fileID: 0}
   _card: {fileID: 0}
   planeOutline: {fileID: 641128984}
-  _Preview: 0
 --- !u!114 &1437247113
 MonoBehaviour:
   m_ObjectHideFlags: 0
