+++ 2D Dogfight Project/Assets/Scripts/Card cs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 549af499090255c45969817597fc80a1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
