+++ 2D Dogfight Project/Assets/Scripts/Card cs/CardAsset.cs
