+++ 2D Dogfight Project/Assets/Scripts/ScriptableObjects cs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cd8b3333741c3a44ea575a4c833adc60
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
