+++ 2D Dogfight Project/Assets/Scripts/ScriptableObjects cs/OnDropEvent.cs.meta
