+++ 2D Dogfight Project/Assets/Scripts/UI cs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a5971bc8345823342b640f7cbbc1247b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
